use `upd_data.json` from neutron-standards-pipeline (0ea6673)
The script `run_pipeline.sh` is used to generate a
maximum likelihood estimate. Then it is used another
time to transfer the evaluated values to the prior
to perform a revised data reduction and transferring
the updated reduced dataset values and prior to
`upd_data.json`. This file replaces `data/data.json`
in this repository and is the starting point for an evaluation.