add GMA database with refined SACS data
The database file is taken from the
`neutron-standards-database` repo where
it is stored under `database/testdata.json`
under commit `6ffabcc`. Currently, the commit
is the tip of the branch labeled
`c3.additional-input+SACS-gmapy`. This state of
the database was also used to produce part of
the results in the paper
R. Capote, G.Schnabel et al,
"Experimental spectrum averaged cross sections (SACS)
in the 252Cf(sf) neutron field and its impact on the evaluation
of neutron standards"
published at
https://doi.org/10.1051/epjconf/202328100027